pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
djongo = "^1.3.6"
google-api-python-client = "^2.23.0"
google-auth-oauthlib = "^0.4.6"
python-decouple = "^3.5"
django-crispy-forms = "^1.13.0"
django-core = "^1.4.1"
beautiful-date = "^2.0.0"
gcsa = "^1.3.0"
[tool.poetry.dev-dependencies]
